+++ packages/genui/lib/src/utils/stream_extensions.dart
@@ -1,7 +1,4 @@
-// Copyright 2025 The Flutter Authors.
-// Use of this source code is governed by a BSD-style license that can be
-// found in the LICENSE file.
-
+import 'dart:async';
 import 'package:stream_transform/stream_transform.dart';
 
 /// Extensions for [Iterable] of [Stream]s.
@@ -15,5 +12,57 @@ extension CombineLatestAll<T> on Iterable<Stream<T>> {
     if (isEmpty) return Stream.value([]);
 
     return first.combineLatestAll(skip(1));
+  }
+}
+
+/// Extensions for [Stream].
+extension SwitchMapExtension<T> on Stream<T> {
+  /// Maps each event to a new stream, and switches to emitting events from
+  /// the most recent inner stream.
+  Stream<R> switchMap<R>(Stream<R> Function(T) convert) {
+    late StreamController<R> controller;
+    StreamSubscription<T>? outerSubscription;
+    StreamSubscription<R>? innerSubscription;
+
+    void cancelInner() {
+      innerSubscription?.cancel();
+      innerSubscription = null;
+    }
+
+    controller = StreamController<R>(
+      sync: true,
+      onListen: () {
+        outerSubscription = listen(
+          (event) {
+            cancelInner();
+            final Stream<R> innerStream = convert(event);
+            innerSubscription = innerStream.listen(
+              (innerEvent) => controller.add(innerEvent),
+              onError: (Object error, StackTrace? stackTrace) =>
+                  controller.addError(error, stackTrace),
+              onDone: () {
+                if (outerSubscription == null) {
+                  controller.close();
+                }
+              },
+            );
+          },
+          onError: (Object error, StackTrace? stackTrace) =>
+              controller.addError(error, stackTrace),
+          onDone: () {
+            if (innerSubscription == null) {
+              controller.close();
+            }
+          },
+        );
+      },
+      onCancel: () {
+        cancelInner();
+        outerSubscription?.cancel();
+        outerSubscription = null;
+      },
+    );
+
+    return controller.stream;
   }
 }